Install Robert Peake's Anti-SpamBot plugin.
Rather than messing with a contact form that seems to only bring in offers to boost my SEO or to outsource development work, I'm moving the site to use a standard email link (after providing other ways to connect). This plugin adds the [email] shortcode, which passes the contents through WordPress' antispambot() function to help obfuscate the address.